fix target editing
